Simplify usage of UserFilterStateMappers
Add static convenience methods to loop through all registered mappers:
- fromDo()
- toDo()

Reorder methods to make code more readable.

456697
